+++ CHANGELOG.md
@@ -9,20 +9,44 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Themes (light and dark);
 - Reference:
   - Added the "Polygon ID State Replication" page;
+  - Added the "World ID State Replication" page;
 - Use Cases:
   - Added the "Identity State Replication" doc;
 - How-to-Guides:
   - Added the Collab.Land integration doc;
+  - Added RariMe guide for issuers;
+  - World ID State Replication setup guide;
 - Contribution guide to the README;
 - Tutorials:
   - Added the Polygon ID Integration usage page;
+- Subscription form (w/o functionality)
+- Mobile sliders
+
 ### Changed
 
+- Design:
+  - Header
+  - Fonts and texts (where needed)
+  - All sections:
+    - Base Layer Section;
+    - Ecosystem Section;
+    - Use Cases Section;
+    - Build Section;
+    - Community (Links) Section;
+    - News Section;
+  - Footer
+  - Docs page design (product based navigation)
 - Updated links from gitlab to github in the whole documentation;
 - Updated the "Overview" doc;
 - Transferred the project to GitHub;
+
+### Fixed
+
+- Minor markup fix in Proof of Humanity tech reference;
+- Minor markup fix in "Setting up validator" guide;
 
 ## [1.6.0] - 28-11-2023
 
@@ -203,7 +227,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Not Found` page returning HTTP status `200`;
 - Broken logo link in the footer;
 
-[Unreleased]: https://github.com/rarimo/docs/compare/v1.5.1...HEAD
+[Unreleased]: https://github.com/rarimo/docs/compare/v1.6.0...HEAD
+[1.6.0]: https://github.com/rarimo/docs/releases/tag/v1.5.1...v1.6.0
 [1.5.1]: https://github.com/rarimo/docs/releases/tag/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/rarimo/docs/releases/tag/v1.4.3...v1.5.0
 [1.4.3]: https://github.com/rarimo/docs/releases/tag/v1.4.2...v1.4.3